+++ src/components/utils/i18n/languages.ts
@@ -23,6 +23,9 @@ const languages = {
     'fileuploader.error.upload.403':
       'Du har ikke rettigheter til å laste opp filer.',
     'fileuploader.error.upload.general': 'Kunne ikke laste opp fil',
+    'searchfield.sr.focus':
+      'Bruk piltaster ned/opp for å velge forslag når de vises',
+    'searchfield.sr.results': '{{ant}} forslag vises.',
     'table.sortable': 'sorterbar',
     'table.sorted_ascending': 'sortert stigende',
     'table.sorted_descending': 'sortert synkende',
@@ -52,6 +55,8 @@ const languages = {
     'fileuploader.error.upload.403':
       'Du har ikkje rettar til å lasta opp filer.',
     'fileuploader.error.upload.general': 'Kunne ikkje lasta opp fil',
+    'searchfield.sr.focus': 'Bruk piltastane ned/opp for å velje',
+    'searchfield.sr.results': '{{ant}} forslag visast.',
     'table.sortable': 'sorterbar',
     'table.sorted_ascending': 'sortert stigande',
     'table.sorted_descending': 'sortert søkkande',
@@ -82,6 +87,8 @@ const languages = {
     'fileuploader.error.upload.403':
       'You do not have permission to upload files',
     'fileuploader.error.upload.general': "Can't upload file",
+    'searchfield.sr.focus': 'Use up/down to choose options when they are shown',
+    'searchfield.sr.results': '{{ant}} items shown.',
     'table.sortable': 'sortable',
     'table.sorted_ascending': 'sorted ascending',
     'table.sorted_descending': 'sorted descending',